let ClassDefCreator dump @@schema_element in multiple-and-indented lines. contains partial fixes for #168.
git-svn-id: http://dev.ctor.org/svn/soap4r/trunk@1653 23c4dcf6-64ea-0310-84d7-d424c7495ec0

+++ test/wsdl/marshal/test_wsdlmarshal.rb
@@ -63,7 +63,7 @@ class TestWSDLMarshal < Test::Unit::TestCase
     gen.opt['force'] = true
     gen.run
     compare("person_org.rb", "Person.rb")
-    File.unlink(pathname('Person.rb'))
+    File.unlink(pathname('Person.rb')) unless $DEBUG
   end
 
   def compare(expected, actual)
